Silence an implicit int-to-float conversion warning on MSVC

+++ Alc/ALu.c
@@ -387,7 +387,7 @@ ALvoid CalcNonAttnSourceParams(ALsource *ALSource, const ALCcontext *ALContext)
     }
     else
     {
-        DryGain *= lerp(1.0f, 1.0f/sqrtf(Device->NumChan), hwidth/F_PI);
+        DryGain *= lerp(1.0f, 1.0f/sqrtf((float)Device->NumChan), hwidth/F_PI);
         for(c = 0;c < num_channels;c++)
         {
             /* Special-case LFE */
